Add accepting or rejecting of XML using errors
Instead of using exceptions for handling the bad XML we now use an
error reference. This can be set from the parser and is read before
acting on each signal received from the signals stream.

The lwt infix operators are removed and replaced with lwt_ppx bindings.
This helps the code look more familiar to actual ocaml syntax whilst
still allowing us to see the lwt sections.

The write_string function now strips the input string and appends a newline.

Some extra logging statements have been added to help with understanding
the behaviour of the unikernel in operation.

A new function for writing error strings to the client has been added.
This just prepends the stripped string with Error to make it clearer.

When XML data is accepted or rejected we write a message to the client
letting them know the outcome.