Re-factor code to not have cyclic dependency
The code with functors was a bit ugly.
I've simplified it so that parsing now returns results which can carry
the messages we need to send out, if any.

This also means the parser module doesn't depend on mirage stack either
which is nicer to have.
Writing tests would be easier since it can just take a stream to parse
from now.