Add pretty printing and stanza creation function
Pretty printing methods are now included with the stanza module. This is
better for expect tests as they don't include the id tag which changes
due to it being a uuid.

The new creator stream_handler is a convenience for the state machine to
generate a new tag to reply to the initial message with.