Add handling of stream features (empty)
The features stanza is now returned if the version of the root stream
tag is as required. Also added tests to capture this response.

Added a function to create this stanza in stanza.ml.

Changed get_attribute_by_name to have an exception variant. This should
be used for when the attribute is expected to be present.

Fixed the test infrastructure in handler, only pulling the first message
from the stream when handling actions.