Add roster set and get
Implemented roster set and get and adapted code to use them.

The roster is replaced with the rosters module which ensures only one
central roster set is available for the server and functions for
updating it are available.

The parser now parses attributes into more abstract representations,
this allows conversion of certain fields to other types such as jid for
from and to.